# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VibeStadium API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones/{zoneId}:
    get:
      operationId: getZone
      tags: [zones]
      summary: Get a single zone by ID
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Zone details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "404":
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /zones/{zoneId}/capacity:
    patch:
      operationId: overrideZoneCapacity
      tags: [ops]
      summary: Override zone capacity (Ops Console)
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapacityOverride"
      responses:
        "200":
          description: Updated zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "404":
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /route:
    get:
      operationId: getRoute
      tags: [route]
      summary: Compute a pathfinding route between zones
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
          description: Origin zone ID
        - name: to
          in: query
          required: true
          schema:
            type: string
          description: Destination zone ID
        - name: stepFreeOnly
          in: query
          required: false
          schema:
            type: boolean
          description: Filter to step-free edges only
      responses:
        "200":
          description: Computed route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResult"
        "400":
          description: Invalid zone IDs or no route found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bins/{binId}/report:
    post:
      operationId: reportBin
      tags: [bins]
      summary: Report a bin as full
      parameters:
        - name: binId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BinReport"
      responses:
        "200":
          description: Bin reported, credits awarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BinReportResult"
        "404":
          description: Bin not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rewards:
    get:
      operationId: getRewards
      tags: [rewards]
      summary: Get fan rewards wallet and transaction ledger
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Rewards wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardsWallet"
  /rewards/claim:
    post:
      operationId: claimReward
      tags: [rewards]
      summary: Claim a reward using credits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardClaim"
      responses:
        "200":
          description: Reward claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardsWallet"
        "400":
          description: Insufficient credits or invalid reward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /narrate:
    post:
      operationId: narrateRoute
      tags: [narrate]
      summary: Get AI narration for a route or zone query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NarrateInput"
      responses:
        "200":
          description: AI narration response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NarrateResult"
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Email already exists or invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current authenticated user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a user's public profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User public profile details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublicProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update the authenticated user's profile details
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (cannot edit other users)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ZoneConnection:
      type: object
      properties:
        toZoneId:
          type: string
        distanceMeters:
          type: number
        hasStairs:
          type: boolean
        hasElevator:
          type: boolean
      required:
        - toZoneId
        - distanceMeters
        - hasStairs
        - hasElevator
    ZoneAccessibility:
      type: object
      properties:
        stepFree:
          type: boolean
        elevatorNearby:
          type: boolean
      required:
        - stepFree
        - elevatorNearby
    Zone:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [gate, concourse, seating, amenity, medical, transit]
        capacityMax:
          type: integer
        capacityCurrent:
          type: number
          description: 0-100 occupancy percentage
        coordinates:
          type: array
          items:
            type: number
          minItems: 2
          maxItems: 2
        connections:
          type: array
          items:
            $ref: "#/components/schemas/ZoneConnection"
        accessibility:
          $ref: "#/components/schemas/ZoneAccessibility"
        trendingDirection:
          type: string
          enum: [rising, falling, stable]
        predictedPct:
          type: ["number", "null"]
          description: Predicted capacity in ~6 minutes
      required:
        - id
        - name
        - type
        - capacityMax
        - capacityCurrent
        - coordinates
        - connections
        - accessibility
        - trendingDirection
    RouteStep:
      type: object
      properties:
        zoneId:
          type: string
        zoneName:
          type: string
        instruction:
          type: string
        distanceMeters:
          type: number
        isElevator:
          type: boolean
      required:
        - zoneId
        - zoneName
        - instruction
        - distanceMeters
        - isElevator
    RouteResult:
      type: object
      properties:
        fromZoneId:
          type: string
        toZoneId:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RouteStep"
        totalDistanceMeters:
          type: number
        estimatedMinutes:
          type: number
        stepFreeOnly:
          type: boolean
        narration:
          type: ["string", "null"]
      required:
        - fromZoneId
        - toZoneId
        - steps
        - totalDistanceMeters
        - estimatedMinutes
        - stepFreeOnly
    Bin:
      type: object
      properties:
        id:
          type: string
        zoneId:
          type: string
        zoneName:
          type: string
        capacityPct:
          type: number
        lastEmptiedAt:
          type: string
        reportCount:
          type: integer
        predictedOverflow:
          type: boolean
      required:
        - id
        - zoneId
        - zoneName
        - capacityPct
        - lastEmptiedAt
        - reportCount
        - predictedOverflow
    BinReport:
      type: object
      properties:
        userId:
          type: string
      required: []
    BinReportResult:
      type: object
      properties:
        bin:
          $ref: "#/components/schemas/Bin"
        creditsAwarded:
          type: integer
        totalCredits:
          type: integer
      required:
        - bin
        - creditsAwarded
        - totalCredits
    RewardTransaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [earned, redeemed]
        amount:
          type: integer
        description:
          type: string
        createdAt:
          type: string
      required:
        - id
        - type
        - amount
        - description
        - createdAt
    RewardOffer:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        cost:
          type: integer
        category:
          type: string
          enum: [food, merch, upgrade, experience]
      required:
        - id
        - title
        - description
        - cost
        - category
    RewardsWallet:
      type: object
      properties:
        userId:
          type: string
        totalCredits:
          type: integer
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/RewardTransaction"
        availableOffers:
          type: array
          items:
            $ref: "#/components/schemas/RewardOffer"
      required:
        - userId
        - totalCredits
        - transactions
        - availableOffers
    RewardClaim:
      type: object
      properties:
        rewardId:
          type: string
        userId:
          type: string
      required:
        - rewardId
...
    NarrateInput:
      type: object
      properties:
        type:
          type: string
          enum: [route, zone, emergency, assistant]
        query:
          type: string
        zoneId:
          type: ["string", "null"]
        routeContext:
          type: ["string", "null"]
          description: JSON-serialized route result for AI context
        language:
          type: string
          default: en
        tone:
          type: string
          enum: [neutral, reassuring, urgent]
          default: neutral
      required:
        - type
        - query
    NarrateResult:
      type: object
      properties:
        text:
          type: string
        urgencyLevel:
          type: string
          enum: [normal, elevated, emergency]
        suggestedZoneId:
          type: ["string", "null"]
        actionLabel:
          type: ["string", "null"]
      required:
        - text
        - urgencyLevel
...
    CapacityOverride:
      type: object
      properties:
        capacityCurrent:
          type: number
          minimum: 0
          maximum: 100
      required:
        - capacityCurrent
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        favoriteTeam:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - createdAt
    UserPublicProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        favoriteTeam:
          type: ["string", "null"]
        createdAt:
          type: string
        credits:
          type: integer
        binsReported:
          type: integer
      required:
        - id
        - name
        - createdAt
        - credits
        - binsReported
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        favoriteTeam:
          type: string
      required:
        - name
        - email
        - password
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        favoriteTeam:
          type: string
        avatarUrl:
          type: string