# lib/api-zod/src/generated/types/routeResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * VibeStadium API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RouteStep } from './routeStep';

export interface RouteResult {
  fromZoneId: string;
  toZoneId: string;
  steps: RouteStep[];
  totalDistanceMeters: number;
  estimatedMinutes: number;
  stepFreeOnly: boolean;
  /** @nullable */
  narration?: string | null;
}