# lib/api-zod/src/generated/types/routeStep.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * VibeStadium API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RouteStep {
  zoneId: string;
  zoneName: string;
  instruction: string;
  distanceMeters: number;
  isElevator: boolean;
}